requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0

# API clients
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.80.0
boto3>=1.28.0

# Image processing
Pillow>=10.0.0

# Async support
aiohttp>=3.8.4
aiobotocore>=2.5.0

# CLI
typer>=0.9.0
rich>=13.3.0

# Web interface (optional)
fastapi>=0.103.0
uvicorn>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
